\name{TraME}

\alias{TraME}
\alias{TraME-package}

\docType{package}

\title{Transportation Methods for Econometrics}

\description{
TraME (Transportation Methods for Econometrics) is an R package for for solving problems of equilibrium computation and estimation in consumer demand and matching frameworks via the Mass Transportation Approach.
}

\details{
\tabular{ll}{
Package: \tab TraME\cr
Type: \tab Package\cr
Version: \tab 0.3.0\cr
Date: \tab 2016-01-14\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab Yes\cr
Depends: \tab rootSolve, Matrix, numDeriv, nloptr, nleqslv, snow, snowfall\cr
Suggests: \tab gurobi\cr}
}

\section{How to cite this package}{

 \preformatted{@Manual{Galichon2016,
 author       = {Alfred Galichon},
 title        = {{TraME}: Transportation Methods for Econometrics},
 year         = {2016},
 note         = {R package version 0.3.0.}}}
\emph{}
}

\section{License}{
TraME is licensed under the GNU General Public License (GPL) version 2, or (at your option) any later version.
}

\author{Alfred Galichon}